+++ src/Atata.HtmlValidation/HtmlValidationOptions.cs
@@ -117,10 +117,10 @@ namespace Atata.HtmlValidation
         /// <summary>
         /// Gets or sets the required version of "html-validate" NPM package.
         /// The required version will be installed if "html-validate" package is not installed or the installed version differs from the required one.
-        /// The default value is <c>"6.7.1"</c>.
+        /// The default value is <c>"7.0.0"</c>.
         /// Set <see langword="null"/> to disable the version check and use any pre-installed version.
         /// </summary>
-        public string HtmlValidatePackageVersion { get; set; } = "6.7.1";
+        public string HtmlValidatePackageVersion { get; set; } = "7.0.0";
 
         /// <summary>
         /// Clones this instance with executing the action that can change some properties of the copy.