+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add optional `AtataContext atataContext` parameter to `HtmlValidator` constructor.
 
 ### Changed
-- Change `HtmlValidationOptions.HtmlValidatePackageVersion` property default value to `"6.7.1"`.
+- Change `HtmlValidationOptions.HtmlValidatePackageVersion` property default value to `"7.0.0"`.
 - Get rid of `AtataContext.Current` usages.
 - Change `HtmlValidationOptions.WorkingDirectoryBuilder` property type from `Func<string>` to `Func<AtataContext, string>`.